+++ src/root.zig
@@ -386,6 +386,7 @@ pub fn stderr() std.Io.File {
 
 test {
     _ = @import("./lang/tests.zig");
+    _ = @import("./ext.zig");
 }
 
 const builtin = @import("builtin");
@@ -424,4 +425,5 @@ pub const async_backend = @import("./vm/runtime/async_backend.zig");
 pub const lang = @import("./lang/root.zig");
 pub const pretty = @import("./pretty.zig");
 pub const std_lib = @import("./std/root.zig");
+pub const ext = @import("./ext.zig");
 pub const std_net = @import("./std/net.zig");